bridge-backend-python/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
aiomysql==0.2.0
python-dotenv==1.0.1
pydantic==2.6.2
cryptography>=42.0.0
